Added switch CROSS_BUILD to organize a bit. Moved platform specific operations to their own files.
