+++ src/r-tech1/dkp/file-system.cpp
@@ -1,0 +1,15 @@
+#ifdef WII
+
+#include "r-tech1/funcs.h"
+#include "r-tech1/file-system.h"
+#include "r-tech1/thread.h"
+#include "r-tech1/system.h"
+#include "r-tech1/utf.h"
+
+// TODO
+using namespace std;
+
+vector<Filesystem::AbsolutePath> Filesystem::getFiles(const AbsolutePath & dataPath, const string & find, bool caseInsensitive){
+}
+
+#endif
