+++ src/r-tech1/dkp/init.cpp
@@ -1,0 +1,3 @@
+#ifdef WII
+#include <fat.h>
+#endif
