+++ src/r-tech1/dkp/system.cpp
@@ -1,0 +1,33 @@
+#if defined(WII)
+
+#include <unistd.h>
+#include <sys/stat.h>
+#include <sys/time.h>
+
+/* devkitpro doesn't have an implementation of access() yet. if it gets one this function
+ * can be removed.
+ */
+int access(const char * path, int mode){
+    struct stat information;
+    int ok = stat(path, &information);
+    // printf("stat of '%s' is %d\n", path.c_str(), ok);
+    if (ok == 0){
+        if (mode == R_OK){
+            if (((information.st_mode & S_IRUSR) == S_IRUSR) ||
+                ((information.st_mode & S_IRGRP) == S_IRGRP) ||
+                ((information.st_mode & S_IROTH) == S_IROTH)){
+                return 0;
+            } else {
+            /* handle other modes if they become useful to us */
+                return -1;
+            }
+       } else {
+           return -1;
+       }
+    } else {
+        // perror("stat");
+        return -1;
+    }
+}
+
+#endif
