+++ src/r-tech1/ps3/file-system.cpp
@@ -1,0 +1,14 @@
+#ifdef PS3
+
+#include "r-tech1/funcs.h"
+#include "r-tech1/file-system.h"
+#include "r-tech1/thread.h"
+#include "r-tech1/system.h"
+#include "r-tech1/utf.h"
+
+// TODO
+using namespace std;
+
+vector<Filesystem::AbsolutePath> Filesystem::getFiles(const AbsolutePath & dataPath, const string & find, bool caseInsensitive){
+}
+#endif
